+++ CHANGELOG.md
@@ -206,3 +206,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Modal component
 - Modal component tests
 
+### Fixed
+
+- Few PropTypes validation errors